+++ src/main/kotlin/io/github/jamalam/extensions/bot/BotUtilityExtension.kt
@@ -1,0 +1,138 @@
+/*
+ * Copyright (C) 2022 Jamalam
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package io.github.jamalam.extensions.bot
+
+import com.kotlindiscord.kord.extensions.extensions.Extension
+import dev.kord.common.entity.PresenceStatus
+import dev.kord.core.Kord
+import io.github.jamalam.api.TopGg
+import io.github.jamalam.util.PRODUCTION
+import io.github.jamalam.util.VERSION
+import io.github.jamalam.util.scheduler
+import kotlinx.coroutines.flow.count
+import kotlinx.datetime.DateTimePeriod
+import kotlin.random.Random
+import kotlin.time.ExperimentalTime
+
+/**
+ * @author  Jamalam360
+ */
+@OptIn(ExperimentalTime::class)
+class BotUtilityExtension : Extension() {
+    override val name = "utility"
+
+    private val presenceDelay = DateTimePeriod(minutes = 2, seconds = 30) //Every 2.5 minutes
+    private val dblDelay = DateTimePeriod(minutes = 30) //Every 10 minutes
+    private val topGg = TopGg()
+
+    override suspend fun setup() {
+        // Set initial presence (without the delay there is an error)
+        scheduler.schedule(10) {
+            this.kord.editPresence {
+                status = PresenceStatus.Idle
+                playing("booting up!")
+            }
+
+            setDBLStats()
+        }
+
+        scheduler.schedule(presenceDelay.seconds.toLong()) {
+            setPresenceStatus()
+        }
+    }
+
+    private suspend fun setPresenceStatus() {
+        BotStatus.random().setPresenceStatus(this.kord)
+        scheduler.schedule(seconds = presenceDelay.seconds.toLong()) {
+            setPresenceStatus()
+        }
+    }
+
+    private suspend fun setDBLStats() {
+        if (PRODUCTION) {
+            topGg.sendServerCount(kord.guilds.count())
+
+            scheduler.schedule(seconds = dblDelay.seconds.toLong()) {
+                setDBLStats()
+            }
+        }
+    }
+}
+
+@Suppress("unused")
+enum class BotStatus(val setPresenceStatus: suspend (kord: Kord) -> Unit) {
+    ServerCount({
+        it.editPresence {
+            status = PresenceStatus.Online
+            watching("over ${it.guilds.count()} servers")
+        }
+    }),
+    CurrentVersion({
+        it.editPresence {
+            status = PresenceStatus.Online
+            playing("Pinguino $VERSION")
+        }
+    }),
+    RandomFun({
+        it.editPresence {
+            status = PresenceStatus.Online
+
+            val status = RandomStatus.random()
+
+            when (status.type) {
+                StatusType.Watching -> watching(status.message)
+                StatusType.Playing -> playing(status.message)
+                StatusType.Listening -> listening(status.message)
+                StatusType.Competing -> competing(status.message)
+            }
+        }
+    });
+
+    companion object {
+        private val values = values()
+        private val size = values.size
+
+        fun random(): BotStatus = values[Random.nextInt(size)]
+    }
+}
+
+@Suppress("unused")
+enum class RandomStatus(val type: StatusType, val message: String) {
+    ListeningForYourCommands(StatusType.Listening, "your commands"),
+    WatchingForYourCommands(StatusType.Watching, "for your commands"),
+    WatchingTheFootball(StatusType.Watching, "the football"),
+    WatchingYou(StatusType.Watching, "you"),
+    PlayingMinecraft(StatusType.Playing, "Minecraft :D"),
+    PlayingABoardGame(StatusType.Playing, "a board game"),
+    CompetingInTheOlympics(StatusType.Competing, "the olympics"),
+    CompetingWithOtherPenguins(StatusType.Competing, "the international penguin tournament");
+
+    companion object {
+        private val values = values()
+        private val size = values.size
+
+        fun random(): RandomStatus = values[Random.nextInt(size)]
+    }
+}
+
+enum class StatusType {
+    Watching,
+    Playing,
+    Listening,
+    Competing
+}