+++ src/main/kotlin/io/github/jamalam/util/Instances.kt
@@ -1,0 +1,42 @@
+/*
+ * Copyright (C) 2022 Jamalam
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package io.github.jamalam.util
+
+import com.kotlindiscord.kord.extensions.utils.scheduling.Scheduler
+import io.github.jamalam.database.Database
+import io.ktor.client.*
+import io.ktor.client.features.json.*
+import io.ktor.client.features.json.serializer.*
+
+/**
+ * @author  Jamalam360
+ */
+
+val scheduler = Scheduler()
+val database = Database()
+val client = HttpClient {
+    install(JsonFeature)
+}
+val lenientClient = HttpClient {
+    install(JsonFeature) {
+        serializer = KotlinxSerializer(json = kotlinx.serialization.json.Json {
+            isLenient = true
+            ignoreUnknownKeys = true
+        })
+    }
+}